backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
python-multipart==0.0.20
python-dotenv==1.1.1
openai==1.97.1
httpx==0.27.0
# requests==2.32.4   # Optional: only keep if you use it somewhere